requirements.txt
fastapi[standard]
line-bot-sdk
pymongo
python-dotenv
opencc-python-reimplemented
sentence_transformers 
ollama
pysqlite3-binary
chromadb
